Overhaul for my personal site
I realized a better way to get people to send me the stuff they want printed, and I can use this as a personal portfolio site. I made this originally like a year ago, and now I get to properly use this.

Some code is AI generated (specifically the email script), this is because I don't care nearly enough to learn an entirely new library for a one off project. I'm going to play a bit more with 98.css to see if I can make my printer site into an iframe that looks like a win98 application.